Move utils to new util file
